Split up `drop_forget_ref` ui test
